Refactoring Collection View and Service concerns
